Reorganise directory structure, and improve class naming
