+++ ecs.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 use PhpCsFixer\Fixer\Import\NoUnusedImportsFixer;
 use PhpCsFixer\Fixer\Operator\NotOperatorWithSuccessorSpaceFixer;
 use PhpCsFixer\Fixer\PhpTag\BlankLineAfterOpeningTagFixer;
+use PhpCsFixer\Fixer\Whitespace\MethodChainingIndentationFixer;
 use Symplify\EasyCodingStandard\Config\ECSConfig;
 use Symplify\EasyCodingStandard\ValueObject\Set\SetList;
 
@@ -16,6 +17,7 @@ return function (ECSConfig $ecsConfig): void {
     $ecsConfig->skip([
         BlankLineAfterOpeningTagFixer::class,
         NotOperatorWithSuccessorSpaceFixer::class,
+        MethodChainingIndentationFixer::class
     ]);
 
     $ecsConfig->rules([
