+++ tests/Attribute/DoctrineTypeMappingTest.php
@@ -1,20 +1,20 @@
 <?php
 declare(strict_types=1);
 
-namespace Headsnet\DoctrineToolsBundle\Tests\CustomTypes;
+namespace Headsnet\DoctrineToolsBundle\Tests\Attribute;
 
-use Headsnet\DoctrineToolsBundle\CustomTypes\CustomType;
+use Headsnet\DoctrineToolsBundle\Attribute\DoctrineTypeMapping;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 
-#[CoversClass(CustomType::class)]
-class CustomTypeTest extends TestCase
+#[CoversClass(DoctrineTypeMapping::class)]
+class DoctrineTypeMappingTest extends TestCase
 {
     #[Test]
     public function name_can_be_specified(): void
     {
-        $sut = new CustomType('custom_name');
+        $sut = new DoctrineTypeMapping('custom_name');
 
         $this->assertEquals('custom_name', $sut->name);
     }
@@ -22,7 +22,7 @@ class CustomTypeTest extends TestCase
     #[Test]
     public function name_is_normalised(): void
     {
-        $sut = new CustomType('some-custom Name');
+        $sut = new DoctrineTypeMapping('some-custom Name');
 
         $this->assertEquals('some_custom_name', $sut->name);
     }
