+++ tests/HeadsnetDoctrineToolsBundleTest.php
@@ -3,9 +3,9 @@ declare(strict_types=1);
 
 namespace Headsnet\DoctrineToolsBundle\Tests;
 
-use Headsnet\DoctrineToolsBundle\CarbonTypes\CarbonTypesCompilerPass;
-use Headsnet\DoctrineToolsBundle\CustomTypes\CustomTypesCompilerPass;
 use Headsnet\DoctrineToolsBundle\HeadsnetDoctrineToolsBundle;
+use Headsnet\DoctrineToolsBundle\Mapping\CarbonTypeMappingsCompilerPass;
+use Headsnet\DoctrineToolsBundle\Mapping\DoctrineTypeMappingsCompilerPass;
 use Nyholm\BundleTest\TestKernel;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
@@ -13,8 +13,8 @@ use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Component\HttpKernel\KernelInterface;
 
 #[CoversClass(HeadsnetDoctrineToolsBundle::class)]
-#[CoversClass(CarbonTypesCompilerPass::class)]
-#[CoversClass(CustomTypesCompilerPass::class)]
+#[CoversClass(CarbonTypeMappingsCompilerPass::class)]
+#[CoversClass(DoctrineTypeMappingsCompilerPass::class)]
 class HeadsnetDoctrineToolsBundleTest extends KernelTestCase
 {
     protected static function getKernelClass(): string
