+++ tests/Mapping/DoctrineTypeMappingNamerTest.php
@@ -1,25 +1,25 @@
 <?php
 declare(strict_types=1);
 
-namespace Headsnet\DoctrineToolsBundle\Tests\CustomTypes;
+namespace Headsnet\DoctrineToolsBundle\Tests\Mapping;
 
-use Headsnet\DoctrineToolsBundle\CustomTypes\CustomTypeNamer;
-use Headsnet\DoctrineToolsBundle\Tests\CustomTypes\Fixtures\DummyCustomType;
-use Headsnet\DoctrineToolsBundle\Tests\CustomTypes\Fixtures\DummyCustomTypeWithName;
+use Headsnet\DoctrineToolsBundle\Mapping\DoctrineTypeMappingNamer;
+use Headsnet\DoctrineToolsBundle\Tests\Mapping\Fixtures\DummyCustomType;
+use Headsnet\DoctrineToolsBundle\Tests\Mapping\Fixtures\DummyCustomTypeWithName;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use ReflectionClass;
 
-#[CoversClass(CustomTypeNamer::class)]
-class CustomTypeNamerTest extends TestCase
+#[CoversClass(DoctrineTypeMappingNamer::class)]
+class DoctrineTypeMappingNamerTest extends TestCase
 {
     #[Test]
     public function default_name_derived_from_class(): void
     {
         $reflection = new ReflectionClass(new DummyCustomType());
 
-        $sut = CustomTypeNamer::getTypeName($reflection);
+        $sut = DoctrineTypeMappingNamer::getTypeName($reflection);
 
         $this->assertEquals('dummy_custom', $sut);
     }
@@ -29,7 +29,7 @@ class CustomTypeNamerTest extends TestCase
     {
         $reflection = new ReflectionClass(new DummyCustomTypeWithName());
 
-        $sut = CustomTypeNamer::getTypeName($reflection);
+        $sut = DoctrineTypeMappingNamer::getTypeName($reflection);
 
         $this->assertEquals('my_custom_name', $sut);
     }
