+++ tests/Mapping/DoctrineTypeMappingsCompilerPassTest.php
@@ -1,20 +1,20 @@
 <?php
 declare(strict_types=1);
 
-namespace Headsnet\DoctrineToolsBundle\Tests\CustomTypes;
+namespace Headsnet\DoctrineToolsBundle\Tests\Mapping;
 
-use Headsnet\DoctrineToolsBundle\CustomTypes\CustomTypeNamer;
-use Headsnet\DoctrineToolsBundle\CustomTypes\CustomTypesCompilerPass;
-use Headsnet\DoctrineToolsBundle\Tests\CustomTypes\Fixtures\DummyCustomType;
-use Headsnet\DoctrineToolsBundle\Tests\CustomTypes\Fixtures\DummyCustomTypeWithName;
+use Headsnet\DoctrineToolsBundle\Mapping\DoctrineTypeMappingNamer;
+use Headsnet\DoctrineToolsBundle\Mapping\DoctrineTypeMappingsCompilerPass;
+use Headsnet\DoctrineToolsBundle\Tests\Mapping\Fixtures\DummyCustomType;
+use Headsnet\DoctrineToolsBundle\Tests\Mapping\Fixtures\DummyCustomTypeWithName;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
-#[CoversClass(CustomTypesCompilerPass::class)]
-#[CoversClass(CustomTypeNamer::class)]
-class RegisterDoctrineTypesCompilerPassTest extends TestCase
+#[CoversClass(DoctrineTypeMappingsCompilerPass::class)]
+#[CoversClass(DoctrineTypeMappingNamer::class)]
+class DoctrineTypeMappingsCompilerPassTest extends TestCase
 {
     #[Test]
     public function can_find_and_register_types(): void
@@ -24,7 +24,7 @@ class RegisterDoctrineTypesCompilerPassTest extends TestCase
         $container->setParameter('headsnet_doctrine_tools.custom_types.scan_dirs', [
             __DIR__ . '/Fixtures',
         ]);
-        $sut = new CustomTypesCompilerPass();
+        $sut = new DoctrineTypeMappingsCompilerPass();
 
         $sut->process($container);
 
@@ -52,7 +52,7 @@ class RegisterDoctrineTypesCompilerPassTest extends TestCase
         $container->setParameter('headsnet_doctrine_tools.custom_types.scan_dirs', [
             __DIR__ . '/Fixtures',
         ]);
-        $sut = new CustomTypesCompilerPass();
+        $sut = new DoctrineTypeMappingsCompilerPass();
 
         $sut->process($container);
